Исправление моделей
